feat: add @override support for users to trump distribution
the behavior of this tool before this commit was to always append
user-created partials to the distribution's.

today, using @override in any partial matching the foliated base
pathing of the distribution's will force the local partial to overwrite
the version in either:

~/.claude/CLAUDE.md
./my-project/AGENTS.md

this also means the canonical trumping and appending use internally by
Claude and Codex continue to have highest cardinality:

./my-project/CLAUDE{,.local}.md
./my-project/AGENTS.override.md

these always trump the distribution AND user-globals as before.

Tests.

a script in test/*.sh is a unit tests that demonstrates the correct

- @override detection
- appending application
- @override application

are all wekk behaved after future refactors.