refactor: replace 'local/' with 'user/'
using 'local/' is misleading, because it implies these alterations are
done relative to a project or current working directory. these files
(./local => ./user) are user-global, not project-local.

    - user's customization hub
    - the kickoff point for new projects
    - fed into every project you work on
    - not generated for a one-project working directory (claude)

this distinction is less obvious and in fact irrelevant for Codex,
because the final built config is the local AGENTS.md. since Claude
takes a direct global -> local system heirarchy, effectivel, this tool
fits in the order unix-like:

system -> user -> local (project)